Troubleshot code, calculator now mostly working
Apply parseInt to storedValue and displayValue so the operator function
works properly; reorder some of the if statements to ensure that
only one condition is met for each time updateDisplay() runs.

Still need to troubleshoot some of the behavior of the '=' key. If user
clicks '=' with a chosen operator & a display value but NO stored value,
the calculator should carry out the operation using the display
value for both 'x' and 'y' and display the result; then if the user
clicks '=' again, the calculator should run operation again with the
current display value and the initial display value (I'm currently using
a new variable tempValue to hold this value, since storedValue can't be
updated for this condition to continue to work) and display the result.

Also need to add 'clear' button.