Solve the click '=' w/o storedValue problem
Solve the problem outlined in the previous commit: now user can click
'=' after clicking an operator and run the operation with the display
value as both x and y in the operate() function; still encountering
an issue with this situation and the 'subtract' operator, though: the
calculator doesn't go down into the negatives if the user keeps clicking
'=', instead it just toggles between 0 and the initial display value.

Delete earlier code that is no longer in use.