Improve display, create 'clear' button, fix bug
Continue tweaking design, including making the font size scalable;
create a 'CLEAR' button that resets the calculator with a display value
of 0; fix a bug where if you clicked equals, then a number, then equals
again, the calculator would perform a calculation using the stored
lastOperator (which is necessary for calculations coming from repeat
clicks of the '=' button). My solution was to add another condition
at the very start of the clickedNumber conditional checking if the last
clicked button was '=' && a lastOperator value exists; if this
condition is met, the calculator clears the lastOperator variable and
proceeds.